added / corrected create multisig tx
when creating a multisig (P2SH) TX, check the target address, and automatically put the OpCodes in the tx. No need to have a special parameter.  
Also fixed a lot in tcls_create.sh, tx fees, presentation of data (utxo amounts,  amount to spend, bitcoinfees.21.co, ...)
And of course the testcase file.