Add files via upload
changed tcls_create.sh to properly wor with the redeem script for the funding tx. At the same time improved tcls_in_sig_script.sh, to properly display signatures and redeem scripts. Then of course the tcls_sign.sh script had to be changed as well, and with this several other files due to dependencies or better represenation of data.